Move argNamedAny to plugin.ArgNamedAny
